refactor: update nativeMessagingVersion filename
