+++ .github/whitelist-capital-letters.txt
@@ -45,7 +45,6 @@
 ./libs/common/src/enums/storageLocation.ts
 ./libs/common/src/enums/themeType.ts
 ./libs/common/src/enums/stateVersion.ts
-./libs/common/src/enums/nativeMessagingVersion.ts
 ./libs/common/src/factories/accountFactory.ts
 ./libs/common/src/factories/globalStateFactory.ts
 ./libs/common/src/factories/stateFactory.ts
