+++ apps/desktop/native-messaging-test-runner/src/commands/bw-credential-retrieval.ts
@@ -3,7 +3,7 @@ import "module-alias/register";
 import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
 
-import { NativeMessagingVersion } from "@bitwarden/common/enums/nativeMessagingVersion";
+import { NativeMessagingVersion } from "@bitwarden/common/enums/native-messaging-version";
 
 import { LogUtils } from "../log-utils";
 import NativeMessageService from "../native-message.service";
