+++ apps/desktop/native-messaging-test-runner/src/commands/bw-status.ts
@@ -1,6 +1,6 @@
 import "module-alias/register";
 
-import { NativeMessagingVersion } from "@bitwarden/common/enums/nativeMessagingVersion";
+import { NativeMessagingVersion } from "@bitwarden/common/enums/native-messaging-version";
 
 import { LogUtils } from "../log-utils";
 import NativeMessageService from "../native-message.service";
