+++ libs/common/src/enums/native-messaging-version.ts
